updated windrose plotting file paths, functions, cleaned up
